requirements.txt
# Core
requests>=2.31.0
click>=8.1.0
# Image handling
Pillow>=10.0.0  # Для мініатюр/обробки зображень
# Msty Studio Integration
pydantic>=2.0.0  # Для валідації конфігів/промптів
pyyaml>=6.0.0    # Якщо Msty використовує YAML для конфігів
websockets>=11.0.0  # Якщо потрібен WebSocket для real-time оновлень
# Logging
python-json-logger>=2.0.7
# Testing
pytest>=7.4.0
responses>=0.23.0
pytest-mock>=3.11.1
mcp>=1.0.0
python-dotenv
# Grok-Api
curl_cffi
fastapi
uvicorn
coincurve
beautifulsoup4
colorama
python-jose[cryptography] 
passlib[bcrypt] 
python-multipart
pysqlite3